+++ baseapp/grpcrouter.go
@@ -1,16 +1,14 @@
 package baseapp
 
 import (
-	"context"
 	"fmt"
 
 	abci "github.com/cometbft/cometbft/api/cometbft/abci/v1"
 	gogogrpc "github.com/cosmos/gogoproto/grpc"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/encoding"
-	"google.golang.org/protobuf/runtime/protoiface"
 
-	"github.com/cosmos/cosmos-sdk/baseapp/internal/protocompat"
+	"github.com/cosmos/cosmos-sdk/baseapp/internal/protoutils"
 	"github.com/cosmos/cosmos-sdk/client/grpc/reflection"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -18,9 +16,9 @@ import (
 )
 
 type QueryRouter interface {
-	HybridHandlerByRequestName(name string) []func(ctx context.Context, req, resp protoiface.MessageV1) error
 	RegisterService(sd *grpc.ServiceDesc, handler interface{})
 	ResponseNameByRequestName(requestName string) string
+	HandlerByRequestName(requestName string) GRPCQueryHandler
 	Route(path string) GRPCQueryHandler
 	SetInterfaceRegistry(interfaceRegistry codectypes.InterfaceRegistry)
 }
@@ -29,9 +27,8 @@ type QueryRouter interface {
 type GRPCQueryRouter struct {
 	// routes maps query handlers used in ABCIQuery.
 	routes map[string]GRPCQueryHandler
-	// hybridHandlers maps the request name to the handler. It is a hybrid handler which seamlessly
-	// handles both gogo and protov2 messages.
-	hybridHandlers map[string][]func(ctx context.Context, req, resp protoiface.MessageV1) error
+	// routesByRequestName maps routes based on the request name
+	routesByRequestName map[string]GRPCQueryHandler
 	// responseByRequestName maps the request name to the response name.
 	responseByRequestName map[string]string
 	// binaryCodec is used to encode/decode binary protobuf messages.
@@ -57,8 +54,11 @@ var (
 func NewGRPCQueryRouter() *GRPCQueryRouter {
 	return &GRPCQueryRouter{
 		routes:                map[string]GRPCQueryHandler{},
-		hybridHandlers:        map[string][]func(ctx context.Context, req, resp protoiface.MessageV1) error{},
+		routesByRequestName:   map[string]GRPCQueryHandler{},
 		responseByRequestName: map[string]string{},
+		binaryCodec:           nil,
+		cdc:                   nil,
+		serviceData:           nil,
 	}
 }
 
@@ -84,12 +84,7 @@ func (qrt *GRPCQueryRouter) Route(path string) GRPCQueryHandler {
 func (qrt *GRPCQueryRouter) RegisterService(sd *grpc.ServiceDesc, handler interface{}) {
 	// adds a top-level query handler based on the gRPC service name
 	for _, method := range sd.Methods {
-		err := qrt.registerABCIQueryHandler(sd, method, handler)
-		if err != nil {
-			panic(err)
-		}
-		err = qrt.registerHybridHandler(sd, method, handler)
-		if err != nil {
+		if err := qrt.registerABCIQueryHandler(sd, method, handler); err != nil {
 			panic(err)
 		}
 	}
@@ -140,35 +135,23 @@ func (qrt *GRPCQueryRouter) registerABCIQueryHandler(sd *grpc.ServiceDesc, metho
 			Value:  resBytes,
 		}, nil
 	}
-	return nil
-}
 
-func (qrt *GRPCQueryRouter) HybridHandlerByRequestName(name string) []func(ctx context.Context, req, resp protoiface.MessageV1) error {
-	return qrt.hybridHandlers[name]
+	// register response name by request name
+	reqName, respName, err := protoutils.RequestAndResponseFullNameFromMethodDesc(sd, method)
+	if err != nil {
+		return fmt.Errorf("grpc router unable to parse request and response name: %w", err)
+	}
+	qrt.responseByRequestName[string(reqName)] = string(respName)
+	qrt.routesByRequestName[string(reqName)] = qrt.routes[fqName]
+	return nil
 }
 
 func (qrt *GRPCQueryRouter) ResponseNameByRequestName(requestName string) string {
 	return qrt.responseByRequestName[requestName]
 }
 
-func (qrt *GRPCQueryRouter) registerHybridHandler(sd *grpc.ServiceDesc, method grpc.MethodDesc, handler interface{}) error {
-	// extract message name from method descriptor
-	inputName, err := protocompat.RequestFullNameFromMethodDesc(sd, method)
-	if err != nil {
-		return err
-	}
-	outputName, err := protocompat.ResponseFullNameFromMethodDesc(sd, method)
-	if err != nil {
-		return err
-	}
-	methodHandler, err := protocompat.MakeHybridHandler(qrt.binaryCodec, sd, method, handler)
-	if err != nil {
-		return err
-	}
-	// map input name to output name
-	qrt.responseByRequestName[string(inputName)] = string(outputName)
-	qrt.hybridHandlers[string(inputName)] = append(qrt.hybridHandlers[string(inputName)], methodHandler)
-	return nil
+func (qrt *GRPCQueryRouter) HandlerByRequestName(requestName string) GRPCQueryHandler {
+	return qrt.routesByRequestName[requestName]
 }
 
 // SetInterfaceRegistry sets the interface registry for the router. This will