+++ x/accounts/defaults/multisig/utils_test.go
@@ -9,7 +9,6 @@ import (
 	types "github.com/cosmos/gogoproto/types/any"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/reflect/protoreflect"
-	"google.golang.org/protobuf/runtime/protoiface"
 
 	"cosmossdk.io/collections"
 	"cosmossdk.io/core/appmodule/v2"
@@ -26,7 +25,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-type ProtoMsg = protoiface.MessageV1
+type ProtoMsg = gogoproto.Message
 
 var TestFunds = sdk.NewCoins(sdk.NewCoin("test", math.NewInt(10)))